refactor: eslint룰에 맞게 수정
