+++ src/components/modal/input_description/renderHandlers.ts
@@ -1,11 +1,12 @@
-import convertHTMLStringToDOM from '../../../utils/convertHTMLStringToDOM';
-import inputDescriptionTemplate from './inputDescriptionTemplate';
-import generateLabelComponent from '../../../uiUtils/generateLabelComponent';
 import DESCRIPTION_LABEL_COMPONENT_DATA from '../../../constants/DescriptionLabelComponentData';
-import generateTextAreaComponent from '../../../uiUtils/generateTextAreaComponent';
-import DESCRIPTION_TEXT_AREA_COMPONENT_DATA from '../../../constants/DescriptionTextAreaComponentData';
-import generateSpanComponent from '../../../uiUtils/generateSpanComponent';
 import DESCRIPTION_SPAN_COMPONENT_DATA from '../../../constants/DescriptionSpanComponentData';
+import DESCRIPTION_TEXT_AREA_COMPONENT_DATA from '../../../constants/DescriptionTextAreaComponentData';
+import generateLabelComponent from '../../../uiUtils/generateLabelComponent';
+import generateSpanComponent from '../../../uiUtils/generateSpanComponent';
+import generateTextAreaComponent from '../../../uiUtils/generateTextAreaComponent';
+import convertHTMLStringToDOM from '../../../utils/convertHTMLStringToDOM';
+
+import inputDescriptionTemplate from './inputDescriptionTemplate';
 
 export const renderBaseDescriptionComponent = (form: Element) => {
   form.appendChild(convertHTMLStringToDOM(inputDescriptionTemplate));
