+++ src/components/modal/select_category/SelectCategory.ts
@@ -1,5 +1,5 @@
 import { categoryChange } from './eventHandlers';
-import { renderSelectCategoryLabelComponent, renderBaseSelectCategoryComponent } from './renderHandlers';
+import { renderBaseSelectCategoryComponent, renderSelectCategoryLabelComponent } from './renderHandlers';
 
 function SelectCategory(form: Element) {
   renderBaseSelectCategoryComponent(form);
