+++ src/components/modal/select_distance/renderHandlers.ts
@@ -1,9 +1,10 @@
-import convertHTMLStringToDOM from '../../../utils/convertHTMLStringToDOM';
-import selectDistanceTemplate from './selectDistanceTemplate';
-import generateLabelComponent from '../../../uiUtils/generateLabelComponent';
 import DISTANCE_LABEL_COMPONENT_DATA from '../../../constants/DistanceLabelComponentData';
 import SELECT_DISTANCE_COMPONENT_DATA from '../../../constants/SelectDistanceComponentData';
+import generateLabelComponent from '../../../uiUtils/generateLabelComponent';
 import generateSelectComponent from '../../../uiUtils/generateSelectComponent';
+import convertHTMLStringToDOM from '../../../utils/convertHTMLStringToDOM';
+
+import selectDistanceTemplate from './selectDistanceTemplate';
 
 export const renderBaseDistanceComponent = (form: Element) => {
   form.appendChild(convertHTMLStringToDOM(selectDistanceTemplate));
