+++ src/components/modal/select_distance/selectDistanceTemplate.ts
@@ -1,3 +1,4 @@
 const selectDistanceTemplate = /*html*/ `
 <div class="form-item form-item--required"></div>`;
+
 export default selectDistanceTemplate;
