+++ src/uiUtils/generateSelectComponent.ts
@@ -3,7 +3,6 @@ import { SelectElementDataType } from '../types/components';
 
 const generateSelectComponent = (selectOptionData: SelectElementDataType) => {
   const { TAG_ID, TAG_NAME, TAG_CLASS_NAME, UI_OPTIONS, TAG_REQUIRED } = selectOptionData;
-
   return Select({
     id: TAG_ID,
     name: TAG_NAME ?? '',
